BURP-10308. Change Version.buildNumber to a long.

+++ src/main/java/burp/api/montoya/core/Version.java
@@ -54,7 +54,7 @@ public interface Version
      *
      * @return The build number.
      */
-    int buildNumber();
+    long buildNumber();
 
     /**
      * The edition of Burp Suite
